refactor(rides): Restructure Ride History screen into modular components
This commit refactors the Wear OS ride history feature to improve code structure and maintainability. The monolithic `RideHistoryPage` has been broken down into smaller, more focused composables: `RideHistoryPageFull`, `RideHistoryCard`, and a simplified `RideHistoryPage`.

This new structure separates concerns, making each component easier to understand, test, and reuse.

- **`RideHistoryPageFull.kt` (New)**:
    - This new file contains the main UI for the ride history screen, a `ScalingLazyColumn` that displays a list of rides.
    - It includes a header, an empty state message, and uses `TitleCard` to display ride details (date, duration, distance, and speed).
    - It also contains a full `@Preview` with mock `BikeRide` data.

- **`RideHistoryCard.kt` (New)**:
    - A new `Card` composable responsible for displaying a single ride entry.
    - It includes the ride date, duration, key stats (distance, average/max speed), and adds a `Delete` button for each ride.

- **`RideHistoryPage.kt` (Refactored)**:
    - The original `RideHistoryPage` has been simplified to act as a container.
    - It now uses a `ScalingLazyColumn` to render a list of the new `RideHistoryCard` components.
    - The detailed UI implementation has been moved to `RideHistoryPageFull.kt` and `RideHistoryCard.kt`.

- **`RideHistoryRoute.kt` (Updated)**:
    - The route has been updated to handle the new `onDeleteClick` action from the `RideHistoryPage`.
    - It now passes the `deleteRide` function from the `RidesViewModel` to the UI layer.
    - The `onRideClick` lambda now correctly passes the `rideId` string for navigation.
